pyproject.toml
[project]
name = "pyOpenSSL-SDK"
dynamic = ["version"]
description = "Wrapper for OpenSSL to create self-signed certificates"
readme = "README.md"
authors = [{ name = "Vignesh Rao", email = "svignesh1793@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Security :: Cryptography"
]
keywords = ["cryptography", "openssl"]
requires-python = ">=3"
dependencies = ["pyOpenSSL"]

[tool.setuptools]
packages = ["OpenSSL_SDK"]

[tool.setuptools.dynamic]
version = {attr = "OpenSSL_SDK.version"}

[build-system]
requires      = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/thevickypedia/pyOpenSSL-SDK"
Source = "https://github.com/thevickypedia/pyOpenSSL-SDK"
"Bug Tracker" = "https://github.com/thevickypedia/pyOpenSSL-SDK/issues"
